requirements.txt
aiogram==2.10.1
aiohttp==3.6.3
aiosqlite==0.15.0
async-timeout==3.0.1
attrs==20.2.0
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
idna==2.10
iso8601==0.1.13
multidict==4.7.6
PyPika==0.42.1
pytz==2020.1
tortoise-orm==0.16.16
typing-extensions==3.7.4.3
yarl==1.5.1
